+++ CHANGELOG.md
@@ -46,6 +46,8 @@ CHANGELOG
 - added `intelmq.bots.experts.mcafee.expert_mar` (1265).
 - renamed `intelmq.bots.experts.ripencc_abuse_contact.expert` to `intelmq.bots.experts.ripe.expert`, compatibility shim will be removed in version 3.0.
   - Added support for geolocation information in ripe expert with a new parameter `query_ripe_stat_geolocation` (#1317).
+  - Restructurize the expert and de-duplicataion (#1384).
+  - Handle '?' in geolocation country data (#1384).
 - `intelmq.bots.experts.ripe.expert`:
   - Use a requests session (#1363).
   - Set the requests parameters once per session.